+++ tomcat-base/src/shared/main/java/eu/xenit/alfresco/tomcat/embedded/Main.java
@@ -6,7 +6,7 @@ import eu.xenit.alfresco.tomcat.embedded.config.DefaultConfigurationProvider;
 import eu.xenit.alfresco.tomcat.embedded.config.EnvironmentVariableConfigurationProvider;
 import eu.xenit.alfresco.tomcat.embedded.share.tomcat.ShareTomcatCustomizer;
 import eu.xenit.alfresco.tomcat.embedded.tomcat.TomcatFactory;
-import eu.xenit.json.jul.JsonFormatter;
+import eu.xenit.logging.json.jul.JsonFormatter;
 import org.apache.catalina.startup.Tomcat;
 
 import java.nio.file.Files;